splitting up contributions in modules
