+++ src/dcmri/contrib_pk.py
@@ -1,8 +1,10 @@
+"""
+Pharmacokinetics
+"""
+
 import math
 import numpy as np
 from scipy.special import gamma
-
-import matplotlib.pyplot as plt
 
 ##### HELPER FUNCTIONS ####
 ###########################
@@ -130,38 +132,6 @@ def conc_nscomp(t, J, K, C0=0):
 
 ######### PACKAGE INTERFACE ##########
 # These functions are exposed to the package user and need to be fully covered by tests, documentation and examples.
-
-def signal_SPGRESS(TR, FA, R1, S0):
-    """Signal aof a spoiled gradient echo sequence in the steady state"""
-    E = np.exp(-TR*R1)
-    cFA = np.cos(FA*math.pi/180)
-    return S0 * (1-E) / (1-cFA*E)
-
-def sample(t, S, ts, dts): 
-    """Sample the signal"""
-    Ss = np.empty(len(ts)) 
-    for k, tk in enumerate(ts):
-        tacq = (t > tk) & (t < tk+dts)
-        Ss[k] = np.average(S[np.nonzero(tacq)[0]])
-    return Ss 
-
-def step_inject(t, weight=70, conc=0.5, dose=0.2, rate=2, start=0):
-    """
-    Injected flux as a steo function
-
-    weight (kg)
-    conc (mmol/mL)
-    dose mL/kg
-    rate mL/sec
-
-    returns dose injected per unit time (mM/sec)"""
-
-    duration = weight*dose/rate     # sec = kg * (mL/kg) / (mL/sec)
-    Jmax = conc*rate                # mmol/sec = (mmol/ml) * (ml/sec)
-    t_inject = (t > start) & (t < start+duration)
-    J = np.zeros(t.size)
-    J[np.nonzero(t_inject)[0]] = Jmax
-    return J
 
 def conc_comp(t, J, K, C0=0):
     """Concentration in a compartment"""
@@ -381,165 +351,5 @@ def flux_ncomp(t, J, K):
     for i in range(nc):
         C[:,i] = K[i,i]*C[:,i]
     return C
-
-def conc_space_1d1c(t, Jp, Jn, Kp, Kn):
-    """Concentration in a spatial 1-compartment model in 1D"""
-    nt = len(Jp)
-    nc = len(Kp)
-    K = Kp + Kn
-    C = np.zeros((nt,nc))
-    for k in range(nt-1):
-        dt = t[k+1]-t[k]
-        # Initialise at current concentration
-        C[k+1,:] = C[k,:]
-        # Add influxes at the boundaries:
-        C[k+1,0] += dt*Jp[k]
-        C[k+1,-1] += dt*Jn[k]
-        # Remove outflux to the neigbours:
-        C[k+1,:] -= dt*K*C[k,:]
-        # Add influx from the neighbours:
-        C[k+1,:-1] += dt*Kn[1:]*C[k,1:]
-        C[k+1,1:] += dt*Kp[:-1]*C[k,:-1]
-    return C
-
-def conc_space_2d1c(t, Jpx, Jnx, Jpy, Jny, Kpx, Knx, Kpy, Kny):
-    """Concentration in a spatial 1-compartment model in 2D"""
-    nt = len(Jpx)
-    nx, ny = Kpx.shape
-    C = np.zeros((nt,nx,ny))
-    K = Kpx + Knx + Kpy + Kny
-    for k in range(nt-1):
-        dt = t[k+1]-t[k]
-        # Initialise at current concentration
-        C[k+1,:,:] = C[k,:,:]
-        # Add influxes at boundaries
-        C[k+1,0,:] += dt*Jpx[k,:]
-        C[k+1,-1,:] += dt*Jnx[k,:]
-        C[k+1,:,0] += dt*Jpy[k,:]
-        C[k+1,:,-1] += dt*Jny[k,:]
-        # Remove outflux to the neigbours
-        C[k+1,:,:] -= dt*K*C[k,:,:]
-        # Add influx from the neighbours
-        C[k+1,:-1,:] += dt*Knx[1:,:]*C[k,1:,:]
-        C[k+1,1:,:] += dt*Kpx[:-1,:]*C[k,:-1,:]
-        C[k+1,:,:-1] += dt*Kny[:,1:]*C[k,:,1:]
-        C[k+1,:,1:] += dt*Kpy[:,:-1]*C[k,:,:-1]
-    return C
-
-def conc_space_1d2c(t, Jp1, Jn1, Jp2, Jn2, Kp1, Kn1, Kp2, Kn2, K12, K21):
-    """Concentration in a spatial 2-compartment model in 1D"""
-    nt = len(Jp1)
-    nc = len(Kp1)
-    K1 = Kp1 + Kn1 + K21
-    K2 = Kp2 + Kn2 + K12
-    C1 = np.zeros((nt,nc))
-    C2 = np.zeros((nt,nc))
-    for k in range(nt-1):
-        dt = t[k+1]-t[k]
-        # Initialise at current concentration
-        C1[k+1,:] = C1[k,:]
-        C2[k+1,:] = C2[k,:]
-        # Add influxes at the boundaries:
-        C1[k+1,0] += dt*Jp1[k]
-        C1[k+1,-1] += dt*Jn1[k]
-        C2[k+1,0] += dt*Jp2[k]
-        C2[k+1,-1] += dt*Jn2[k]
-        # Remove outflux to the neigbours:
-        C1[k+1,:] -= dt*K1*C1[k,:]
-        C2[k+1,:] -= dt*K2*C2[k,:]
-        # Add influx from the neighbours:
-        C1[k+1,:-1] += dt*Kn1[1:]*C1[k,1:]
-        C1[k+1,1:] += dt*Kp1[:-1]*C1[k,:-1]
-        C2[k+1,:-1] += dt*Kn2[1:]*C2[k,1:]
-        C2[k+1,1:] += dt*Kp2[:-1]*C2[k,:-1]
-        # Add influx at the same location
-        C1[k+1,:,:] += dt*K21*C1[k,:,:]
-        C2[k+1,:,:] += dt*K12*C2[k,:,:]
-    return C1, C2
-
-def conc_space_2d2c(t, 
-            Jp1x, Jn1x, Jp2x, Jn2x, 
-            Jp1y, Jn1y, Jp2y, Jn2y, 
-            Kp1x, Kn1x, Kp2x, Kn2x, 
-            Kp1y, Kn1y, Kp2y, Kn2y, 
-            K12, K21):
-    """Concentration in a spatial 2-compartment model in 2D"""
-    nt = len(Jp1x)
-    nx, ny = Kp1x.shape
-    C = np.zeros((nt,nx,ny))
-    K1 = Kp1x + Kn1x + Kp1y + Kn1y + K21
-    K2 = Kp2x + Kn2x + Kp2y + Kn2y + K12
-    C1 = np.zeros((nt,nx,ny))
-    C2 = np.zeros((nt,nx,ny))
-    for k in range(nt-1):
-        dt = t[k+1]-t[k]
-        # Initialise at current concentration
-        C1[k+1,:,:] = C1[k,:,:]
-        C2[k+1,:,:] = C2[k,:,:]
-        # Add influxes at the boundaries:
-        C1[k+1,0,:] += dt*Jp1x[k,:]
-        C1[k+1,-1,:] += dt*Jn1x[k,:]
-        C1[k+1,:,0] += dt*Jp1y[k,:]
-        C1[k+1,:,-1] += dt*Jn1y[k,:]
-        C2[k+1,0,:] += dt*Jp2x[k,:]
-        C2[k+1,-1,:] += dt*Jn2x[k,:]
-        C2[k+1,:,0] += dt*Jp2y[k,:]
-        C2[k+1,:,-1] += dt*Jn2y[k,:]
-        # Remove outflux to the neigbours:
-        C1[k+1,:,:] -= dt*K1*C1[k,:,:]
-        C2[k+1,:,:] -= dt*K2*C2[k,:,:]
-        # Add influx from the neighbours
-        C1[k+1,:-1,:] += dt*Kn1x[1:,:]*C1[k,1:,:]
-        C1[k+1,1:,:] += dt*Kp1x[:-1,:]*C1[k,:-1,:]
-        C1[k+1,:,:-1] += dt*Kn1y[:,1:]*C1[k,:,1:]
-        C1[k+1,:,1:] += dt*Kp1y[:,:-1]*C1[k,:,:-1]
-        C2[k+1,:-1,:] += dt*Kn2x[1:,:]*C2[k,1:,:]
-        C2[k+1,1:,:] += dt*Kp2x[:-1,:]*C2[k,:-1,:]
-        C2[k+1,:,:-1] += dt*Kn2y[:,1:]*C2[k,:,1:]
-        C2[k+1,:,1:] += dt*Kp2y[:,:-1]*C2[k,:,:-1]
-        # Add influx at the same location
-        C1[k+1,:,:] += dt*K21*C1[k,:,:]
-        C2[k+1,:,:] += dt*K12*C2[k,:,:]
-    return C1, C2
-
-
-        
-
-######### TESTS ############
-
-tmax = 120 # sec
-dt = 0.01 # sec
-MTT = 20 # sec
-
-weight = 70.0           # Patient weight in kg
-conc = 0.25             # mmol/mL (https://www.bayer.com/sites/default/files/2020-11/primovist-pm-en.pdf)
-dose = 0.025            # mL per kg bodyweight (quarter dose)
-rate = 1                # Injection rate (mL/sec)
-start = 20.0        # sec
-dispersion = 90    # %
-
-
-def test_step_inject():
-    t = np.arange(0, tmax, dt)
-    J = step_inject(t, weight, conc, dose, rate, start)
-    assert np.round(np.sum(J)*dt,1) == np.round(weight*dose*conc, 1)
-
-
-if __name__ == "__main__":
-
-    test_step_inject()
-
-
-
-
-
-
-
-
-
-
-
-
-
